I complete all CRUD method of user, tag, todo, and brash up the request format. The rest of things to do is something around authentication.

+++ application/tag/tag_delete_use_case.go
@@ -17,7 +17,7 @@ func NewDeleteTagUseCase(tagRepository tagDomain.TagRepository) *DeleteTagUseCas
 // todo項目削除
 type DeleteTagUseCaseInputDto struct {
 	Id    string
-	TagId uint
+	TagId uint64
 }
 
 // 特定の項目を削除する
